# src/main/resources/static/swagger/index.yaml
swagger: '2.0'
info:
  description: 盘锦酒店分销系统 V1.0 API接口文档
  version: 1.0.0
  title: 盘锦酒店分销
basePath: /panjin_jiudian
schemes:
  - http
securityDefinitions:
  api_key:
    type: apiKey
    name: token
    in: header
paths:
...
  /app/startPage:
    get:
      tags:
        - 首页
      summary: 获取首页信息
      produces:
        - application/json
      parameters:
        - name: roomPage
          description: 房间当前页
          in: query
          type: integer
          default: 1
          required: true
        - name: roomLimit
          description: 房间每页条数
          in: query
          type: integer
          default: 5
          required: true
        - name: goodPage
          description: 商品当前页
          in: query
          type: integer
          default: 1
          required: true
        - name: goodLimit
          description: 商品每页条数
          in: query
          type: integer
          default: 5
          required: true
        - name: cmsPage
          description: 新闻当前页
          in: query
          type: integer
          default: 1
          required: true
        - name: cmsLimit
          description: 新闻每页条数
          in: query
          type: integer
          default: 5
          required: true
        - name: categoryId
          description: 商品categoryId
          in: query
          type: integer
          required: true
      responses:
        '200':
          schema:
             $ref: '#/definitions/R'
...
  /app/getInvoiceDetail:
    get:
      tags:
        - 发票管理
      summary: 获取发票详细信息
      produces:
        - application/json
      parameters:
        - name: invoiceId
          description: 发票ID
          in: query
          type: integer
          required: true
      responses:
        '200':
          schema:
             $ref: '#/definitions/R'
  /app/addInvoice:
    post:
      tags:
        - 发票管理
      summary: 新增发票配置
      produces:
        - application/json
      parameters:
         - name: body
           description: json对象
           in: body
           type: string
           schema: 
              $ref: '#/definitions/InvoiceForm'
           required: true
      responses:
        '200':
          schema:
             $ref: '#/definitions/R'
  /app/dropInvoice:
    post:
      tags:
        - 发票管理
      summary: 删除发票配置
      produces:
        - application/json
      parameters:
        - name: invoiceId
          description: 发票ID
          in: query
          type: integer
          required: true
      responses:
        '200':
          schema:
             $ref: '#/definitions/R'
...
  /app/login:
    post:
      tags:
        - 用户登录
      summary: 用户登录
      produces:
        - application/json
      parameters:
        - name: body
          description: 管理员对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/LoginAppForm'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/Login'
...
  /app/userAgreement:
    get:
      tags:
        - 用户注册
      summary: 获取协议文档
      produces:
        - application/json
      parameters:
        - name: type
          description: 文档类型
          in: query
          type: integer
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
definitions:
  R:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      msg:
        description: 失败原因
        type: string
...
  Login:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      token:
        description: token
        type: string
      expire:
        description: 过期时长
        type: integer
        format: int32
      msg:
        description: 失败原因
        type: string
...
  LoginAppForm:
    type: object
    properties:
      mobile:
        description: 手机号
        type: string
        default: 13677375913
      password:
        description: 密码
        type: string
        default: program
...
  InvoiceForm:
    type: object
    properties:
      invoiceId:
        description: 发票ID（0为新增发票）
        type: integer
      isCpmpany:
        description: 是否公司发票0公司1个人
        type: integer
      invoiceType:
        description: 发票类型0增值税普票1增值税专票
        type: integer
      companyName:
        description: 企业名称（个人发票时为开票人姓名）
        type: string
      taxpayerCode:
        description: 纳税人识别编号
        type: string
      socialCreditCode:
        description: 统一社会信用代码
        type: string
      companyAddress:
        description: 公司注册地址
        type: string
      telNum:
        description: 联系电话
        type: string
      bank:
        description: 对外付款开户行
        type: string
      bankAccount:
        description: 对外付款银行账号
        type: string
...